rainforest: avoid the slow memcpy() operation to reset the state
It's really expensive to use memcpy() to copy 16 kB of data on some
small processors like ARM Cortex A53 which only have 64-bit data paths
from the L1 cache. This roughly consumes 2k cycles just for the copy.
"Perf top" shows that half of the time is spent in memcpy(), and given
that this exhausts the L1 cache, the rest of the operations must cause
a lot of thrashing.

Since there are few modifications applied to the rambox between two
consecutive calls, better keep a history of recent changes inside
the context itself. This doesn't cost much because the write bus
between the CPU and the L1 cache is 128 bit on A53 so we can afford
a few writes. Also, the typical amount of updates apparently is
between 16 and 32 so it makes sense to put an upper bound on 32 and
remain the memory footprint low..

The performance is roughly multiplied by 5 on A53 just by doing this,
the hash rate reaches about 14.4k/s on NanoPI-Neo4, or almost 10 times
the performance of the original code.