rf_crc32: split rf_add64_crc32() in two
This function has one part which is standard CRC and another part
which is for the RF algo. Let's keep only the standard part in the
CRC. This allows to simplify it not to perform shift operations on
64 bit registers when using the table since 32 are sufficient. This
was tested both on x86_64 and arm64. The x86_64 code is now 1% faster.