implement a variable number of loops
this idea was first suggested by @wtarreau on github :
   https://github.com/bschn2/rainforest/pull/7#issuecomment-476403614

It was a bit refined by exponentiating the sine wave with an odd power, providing a heartbeat-like curve giving short variations around a moderately stable average. The max number of rounds can now reach 765 with an average around 383.

The sine is passed a 32 bit discrete value and its output is rounded to an integer number of loops as well. IEEE754-compliant implementations return the exact same values when operating on 64 bit floats. Simplified implementations and those working based on successive approximations will get a number of values wrong possibly even the remainder of the input. A wrong number of rounds will result in a wrong hash.

The OpenCL implementation was validated against the x86_64 one and both return similar hashes.