refine the sin_scaled function to pack extremes more
In this discussion https://github.com/bschn2/rainforest/issues/15 @wtarreau experimented with two sqrt() around the sine to pack values better. This experiment proves to provide smoother oscillations and can be simplified by reducing the exponentiation by two. Also the extra addition is simplified with a call to round(). With this we can safely divide the number of rounds by almost 3 and see the hash rate increase by as much. This needs to adjust the memory access ratios however (4 times more) which overall increases both performance and security.