rfv2_core: avoid a useless AND operation in rotates
Clang emits a useless AND operation on the shift count in the rotate
operations on ARM to zero-extend the count. This is not needed here,
let's simply make the bit shift count the same type as the value so
that no operation is needed. It is also very slightly faster with
clang on x86_64.