rfv2_test: properly reset the sigalarm handler
When built on i586, only the first line is emitted and "Alarm clock" is
displayed and aborts the signal. This is because the signal is automatically
unset. We only need to set it again to its handler.

On i586 we can see exactly half of the single-core performance the machine
delivers in 64-bit and slightly less than half the performance with 8 threads.