rfv2_core: implement read-only rambox support
With only 2 loops I'm seeing 180 rambox accesses per hash on average,
which gives an extremely low probability of meeting a modified value
on 12M entries, typically 1 hash out of 70000 will have this. Since we
don't care about all solutions and can affort to miss some good ones
and to have a few bad ones from time to time, here we implement a mode
consisting in not updating the rambox during the scans. A few hashes
will be wrong from time to time but will be detected by the caller's
check. This will slightly increase the performance by not writing to
memory and will allow to use a unified rambox in some circumstances.