rfv2_core: move the hash verification to rfv2_scan_hdr()
It's not correct to return invalid hashes and let the caller deal with
them. Actually it's much easier that rfv2_scan_hdr() performs a full
hash when it believes it found a solution and only returns in this case.
It makes the caller not have to deal with these few invalid hashes. The
cpuminer code was updated to drop this test.