rfv2_test: make the bench code actually be a copy of rfv2_scan_hdr()
The purpose of the bench is to see how fast the scan function works,
so it should perform the same operations and report the speed in this
situation. Here we make run_bench() a copy of this function except that
it doesn't have a min/max for the nonces, has no target nor output so
it runs continuously. The reported numbers are around 5% higher than
what the previous version used to report.