rfv2_cpuminer: make the scan function return the number of hashes done
This way we can have cpuminer report a correct hash rate. However cpuminer
also uses this to calculate how far to go with the next nonce, so we need
to adjust the max number of nonces checked in one round to avoid looping
too fast, hence the cpuminer patch was updated. Now rfv2_test and cpuminer
report the same performance level.